requirements.txt
seaborn >=0.11.2
statsmodels >=0.13.2
scikit-learn >=1.0.2
requests >=2.27.1
# jupyterlab>=3.0.0
pytest>=7.1.1
pytest-cov>=3.0.0
pytest-mock>=3.7.0
pytest-mpl>=0.14
pre-commit>=2.18.1
psycopg2-binary>=2.9.3
python-dotenv>=0.20.0
lxml>=4.8.0
networkx>=3.1
session-info>=1.0.0
pingouin>=0.5.3
gglasso>=0.1.9
ipykernel>=6.22.0
miceforest>=5.6.3
